+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * Fix Api.get_register\_data\ returning wrong object
   * Fix check of :num\_scouts in Section.initalize
   * Fix check of :section\_id in Member.initalize (apparently 0 is allowd in the API return data)
+  * Fix role's section not being set from API data
 
 ## Version 0.0.13